refactor!: move build logic to its own file and only export the build function
